feat: crud rooms

+++ src/app/pages/rooms/main-room/rooms.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Rooms } from 'src/app/shared/interfaces/rooms/rooms.interface';
+import { Router } from '@angular/router';
+import { Room } from 'src/app/shared/interfaces/rooms/rooms.interface';
 import { RoomsService } from 'src/app/shared/services/rooms.service';
+import { SweetalertService } from 'src/app/shared/services/sweetalert.service';
 
 @Component({
   selector: 'app-rooms',
@@ -8,7 +10,8 @@ import { RoomsService } from 'src/app/shared/services/rooms.service';
   styleUrls: ['./rooms.component.css'],
 })
 export class RoomsComponent implements OnInit {
-  rooms: Rooms[] = [];
+  hasContent: boolean = false;
+  rooms: Room[] = [];
 
   displayedColumns: string[] = [
     'roomNumber',
@@ -19,7 +22,11 @@ export class RoomsComponent implements OnInit {
     'actions',
   ];
 
-  constructor(private roomsService: RoomsService) {}
+  constructor(
+    private roomsService: RoomsService,
+    private router: Router,
+    private sweetalertService: SweetalertService
+  ) {}
 
   ngOnInit(): void {
     this.getAllRooms();
@@ -27,13 +34,23 @@ export class RoomsComponent implements OnInit {
 
   getAllRooms(): void {
     this.roomsService.getAllRooms().subscribe((response) => {
+      this.hasContent = response.length > 0;
       this.rooms = response;
     });
   }
 
   deleteRoom(id: number): void {
-    this.roomsService.deleteRoom(id).subscribe((response) => {
-      this.getAllRooms();
+    this.sweetalertService.confirmAlert().then((result) => {
+      if (result.isConfirmed) {
+        this.roomsService.deleteRoom(id).subscribe((response) => {
+          this.getAllRooms();
+        });
+        this.sweetalertService.simpleAlert('Hotel eliminado');
+      }
     });
+  }
+
+  public setRoomId(id: number): void {
+    this.router.navigateByUrl(`rooms/form?id=${id}`);
   }
 }
